refac: consolidate chat logic into single component
